fix: Errors with missing posix extension (#7666)
* fix: errors with missing posix extension
It should be an optional dependency

* add php71w-process to the CentOS 7 install docs

+++ LibreNMS/Proc.php
@@ -197,7 +197,10 @@ class Proc
 
         if (!$closed) {
             // try harder
-            $killed = posix_kill($status['pid'], 9); //9 is the SIGKILL signal
+            $killed = false;
+            if (function_exists('posix_kill')) {
+                $killed = posix_kill($status['pid'], 9); //9 is the SIGKILL signal
+            }
             proc_close($this->_process);
 
             if (!$killed && $this->isRunning()) {
